feat(main) 대댓글 완성, 내일 ppt 만들고 이력서 넣고 남은 시간에 회원가입 로그인 만들기

+++ just_board_api/src/main/java/com/just_board/just_board/entity/Comment.java
@@ -37,11 +37,11 @@ public class Comment {
     // private Long articleId;
     private Date date;
     @Column(name="ref_id")
-    private Long refId;
-    private Long step;
-    private Long ref;
+    private int refId;
+    private int step;
+    private int ref;
     @Column(name="ref_order")
-    private Long refOrder;
+    private int refOrder;
     @Column(name="answer_num")
     private Long answerNum;    
 }
